diffusert/requirements.txt
diffusers
accelerate
transformers
xformers
ninja
scipy
aiohttp
aiohttp_cors
av
aiortc
torchvision
ray
langchain
langchain_experimental
langserve
uvicorn
fastapi
sse_starlette
pydantic==1.10.9